fix: add no-ssr to component
