+++ app/components/FormCreator/FormBlocks/index.js
@@ -1,27 +1,47 @@
 import { Draggable, Droppable } from 'react-beautiful-dnd';
+import { getRenderItem } from '../_utils'
 
-export default function FormBlocks(props) {
+function Copyable(props) {
   return (
-    <Droppable droppableId="BAG">
+    <Droppable
+      renderClone={getRenderItem(props.items, props.className)}
+      droppableId={props.droppableId}
+      isDropDisabled={true}
+    >
       {(provided, snapshot) => (
-        <ul ref={provided.innerRef} className="shopping-bag">
-          {props.items.map((item, index) => (
-            <Draggable key={item.id} draggableId={item.id} index={index}>
-              {(provided, snapshot) => (
-                <li
-                  ref={provided.innerRef}
-                  {...provided.draggableProps}
-                  {...provided.dragHandleProps}
-                  style={provided.draggableProps.style}
-                >
-                  {item.label}
-                </li>
-              )}
-            </Draggable>
-          ))}
+        <ul ref={provided.innerRef} className={props.className}>
+          {props.items.map((item, index) => {
+            const shouldRenderClone = item.id === snapshot.draggingFromThisWith;
+            return (
+              <React.Fragment key={item.id}>
+                {shouldRenderClone ? (
+                  <li className="react-beatiful-dnd-copy">{item.label}</li>
+                ) : (
+                  <Draggable draggableId={item.id} index={index}>
+                    {(provided, snapshot) => (
+                      <React.Fragment>
+                        <li
+                          ref={provided.innerRef}
+                          {...provided.draggableProps}
+                          {...provided.dragHandleProps}
+                          className={snapshot.isDragging ? "dragging" : ""}
+                        >
+                          {item.label}
+                        </li>
+                      </React.Fragment>
+                    )}
+                  </Draggable>
+                )}
+              </React.Fragment>
+            );
+          })}
           {provided.placeholder}
         </ul>
       )}
     </Droppable>
   );
+}
+
+export default function Shop(props) {
+  return <Copyable droppableId="SHOP" className="shop" items={props.items} />;
 }
