+++ pages/_document.js
@@ -53,7 +53,7 @@ MyDocument.getInitialProps = async (ctx) => {
   // 4. page.render
 
   // Render app and page and get the context of the page with collected side effects.
-  resetServerContext();
+  // resetServerContext();
   const sheets = new ServerStyleSheets();
   const originalRenderPage = ctx.renderPage;
